refactor: move endpoint classes to particular package
