+++ lib/src/main/java/hajimeapi4j/internal/datatype/utilizations/Music.java
@@ -1,8 +1,8 @@
 package hajimeapi4j.internal.datatype.utilizations;
 
-import hajimeapi4j.internal.EndPointImpl;
 import hajimeapi4j.internal.datatype.Member;
 import hajimeapi4j.internal.datatype.Unit;
+import hajimeapi4j.internal.endpoint.EndPointImpl;
 import java.util.List;
 import java.util.Optional;
 import javax.annotation.Nonnull;
