+++ lib/src/main/java/hajimeapi4j/internal/endpoint/EndPointImpl.java
@@ -1,4 +1,4 @@
-package hajimeapi4j.internal;
+package hajimeapi4j.internal.endpoint;
 
 import hajimeapi4j.api.endpoint.EndPoint;
 import hajimeapi4j.api.request.RestAction;
