+++ lib/src/main/java/hajimeapi4j/internal/endpoint/TaxEndPointImpl.java
@@ -1,4 +1,4 @@
-package hajimeapi4j.internal;
+package hajimeapi4j.internal.endpoint;
 
 import hajimeapi4j.api.endpoint.EndPoint;
 import hajimeapi4j.api.endpoint.TaxEndPoint;
